cpufreq: interactive: call __cpufreq_driver_target() for cur frequency
__cpufreq_driver_target() is not getting called by the governor
when the current frequency is within the new policy limits when some
external client (say thermal mitigation) change the policy limits.
However, there is a race here that current frequency might be
getting set in parallel by speedchange task which can be higher than
the new policy->max which was set by external client. This results the
current frequency being higher than the policy->max and voilating the
policy limits. This causes thermal impact as well and in turn high power
consumption. So Fix this by calling __cpufreq_driver_target() even for
the current frequency and update the cpu frequency inside the cpufreq
driver based on the policy check.

Change-Id: I3ca9010a2b5ad93f6d9a5c02b227ea9d98b639f3
Signed-off-by: Hanumath Prasad <hpprasad@codeaurora.org>